testing/test-api-performance-testing-mcp/infrastructure-cdk/requirements.txt
aws-cdk-lib>=2.176.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha>=2.176.0a0
cdk-nag>=2.37.0